BASE_URL constant was created in the chared folder, created interceptors in the lib folder to check the request and respond to errors for different url's,  the query was moved to the api folder in the customers directory,  and a folder for table columns was created
