+++ src/pages/customers/index.ts
@@ -1,4 +1,5 @@
 import Customers from './ui/customers.tsx'
 import usePaginationStore from './model/store.ts'
+import CustomersColumns from './ui/customers-columns.tsx'
 
-export { Customers, usePaginationStore }
+export { Customers, usePaginationStore, CustomersColumns }
