+++ src/pages/customers/ui/customers-columns.tsx
@@ -1,0 +1,52 @@
+import { Avatar, AvatarProps, Button } from 'antd'
+import { UserSwitchOutlined, EditOutlined } from '@ant-design/icons'
+import { ColumnsType } from 'antd/es/table'
+
+type Customer = {
+  id: string
+  name: string
+  email: string
+  phone: number
+  address: string
+  avatar: string
+  avatarProps?: AvatarProps
+}
+
+const CustomersColumns = (): ColumnsType<Customer> => [
+  {
+    title: <UserSwitchOutlined />,
+    dataIndex: 'avatar',
+    key: 'avatar',
+    render: (_, { avatar, avatarProps }) => {
+      return <Avatar size="large" src={avatar} {...avatarProps} />
+    },
+  },
+  {
+    title: 'Name',
+    dataIndex: 'name',
+    key: 'name',
+  },
+  {
+    title: 'Email',
+    dataIndex: 'email',
+    key: 'email',
+  },
+  {
+    title: 'Phone',
+    dataIndex: 'phone',
+    key: 'phone',
+  },
+  {
+    title: 'Address',
+    dataIndex: 'address',
+    key: 'address',
+  },
+  {
+    title: 'Edit',
+    dataIndex: 'edit',
+    key: 'edit',
+    render: () => <Button icon={<EditOutlined />} />,
+  },
+]
+
+export default CustomersColumns
