+++ src/shared/constants.ts
@@ -1,0 +1,2 @@
+const BASE_URL = 'http://localhost:3001'
+export default BASE_URL
